refactor: update dropdown menu to be a more generic dropdown
